requirements.txt
Flask
mysql-connector-python
Flask-SQLAlchemy
flask-session
requests